Implement overCode for markdown files
