+++ src/languages.ts
@@ -1,1 +1,3 @@
-export type Lanuage = "idris" | "lidr" | "markdown"
+export type ExtLanguage = "idris" | "lidr" | "markdown"
+
+export const isExtLanguage = (s: string): s is ExtLanguage => ["idris", "lidr", "markdown"].includes(s)
