+++ src/state.ts
@@ -1,7 +1,7 @@
 import { ChildProcess, spawn } from "child_process"
 import { IdrisClient, Reply } from "idris-ide-client"
 import * as vscode from "vscode"
-import { Lanuage } from "./languages"
+import { ExtLanguage } from "./languages"
 import { handleWarning } from "./providers/diagnostics"
 import { VirtualDocInfo } from "./providers/virtual-docs"
 
@@ -86,5 +86,5 @@ export const initialiseState = () => {
   if (hoverAction) state.hoverAction = hoverAction
 }
 
-export const supportedLanguages = (state: State): Lanuage[] =>
+export const supportedLanguages = (state: State): ExtLanguage[] =>
   state.idris2Mode ? ["idris", "lidr", "markdown"] : ["idris", "lidr"]
